feat(timestamp): standardize timestamp format to i64 milliseconds across all traits and types
- Updated Account trait to use i64 timestamps instead of u64 with comprehensive documentation
- Added deprecated u64 wrapper methods for backward compatibility in Account and Funding traits
- Updated Funding trait with i64 timestamp format and detailed migration examples
- Modified all timestamp fields in funding rate related types to use i64 instead of u64
- Updated Binance exchange implementation to use i64 timestamps in fetch_my_trades and related methods
- Updated example files to demonstrate proper i64 timestamp usage throughout the library
- Enhanced documentation with timestamp format specifications and usage examples
- Added comprehensive timestamp format documentation to all relevant traits
- Updated Binance futures implementation to use i64 for funding rate history methods
- Fixed test cases to properly handle i64 timestamps and removed ignore annotations
- Added timestamp utility demonstrations in example files
- Updated Cargo.toml files with new dependencies and benchmark configurations

+++ ccxt-exchanges/src/binance/ws.rs
@@ -2377,7 +2377,7 @@ impl BinanceWs {
             percentage: None,
             position_side: None,
             dual_side_position: None,
-            timestamp: Some(chrono::Utc::now().timestamp_millis() as u64),
+            timestamp: Some(chrono::Utc::now().timestamp_millis()),
             datetime: Some(chrono::Utc::now().to_rfc3339()),
         })
     }
